Move all Pilcrow extensions to shared file
