+++ Demo/macOS/App/AppDelegate.swift
@@ -1,18 +1,7 @@
-//
-//  AppDelegate.swift
-//  Demo macOS
-//
-//  Created by Zach Waugh on 11/11/21.
-//
-
 import Cocoa
 
 @main
 class AppDelegate: NSObject, NSApplicationDelegate {
-
-    
-
-
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         // Insert code here to initialize your application
     }
@@ -24,7 +13,4 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationSupportsSecureRestorableState(_ app: NSApplication) -> Bool {
         return true
     }
-
-
 }
-
